% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_email.R
\name{b2n_add_subscribers}
\alias{b2n_add_subscribers}
\title{add matching subscribers}
\usage{
b2n_add_subscribers(post_email, subscriber_data = NULL)
}
\arguments{
\item{post_email}{a blog post email created by \code{b2n_post_to_email}}

\item{subscriber_data}{subscriber data from \code{b2n_fetch_subscribers}}
}
\value{
list
}
\description{
Given a blog post email and subscriber data, finds subscribers who have matching categories,
and adds them to the bcc field of the email, if any exist.
}
