% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2n_subscribers.R
\name{b2n_clean_subscribers}
\alias{b2n_clean_subscribers}
\title{clean subscriber file}
\usage{
b2n_clean_subscribers(subscriber_info)
}
\arguments{
\item{subscriber_info}{a data.frame from reading the Google Sheet}
}
\value{
tibble
}
\description{
Given fetched subscriber Google Sheet, cleans it and reformats it to be more useful.
}
