% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_email.R
\name{b2n_compose}
\alias{b2n_compose}
\title{composes the email}
\usage{
b2n_compose(
  blog_dir,
  post = "latest",
  subscriber_data = NULL,
  from_email = "",
  extra_subject = ""
)
}
\arguments{
\item{blog_dir}{where the blog is (see details)}

\item{post}{whether to fetch the \strong{latest} post or matching a title}

\item{subscriber_data}{the previously fetched subscriber data}

\item{from_email}{who is sending the email}

\item{extra_subject}{any extra text to add to the email subject}
}
\value{
gmailr html message
}
\description{
Given the blog directory, finds the RSS feed file, and grabs either the \strong{latest}
or the post with a title matching the supplied title, and then finds the subscribers
that have the matching categories as in the post.
}
\details{
While it is expected that \strong{most} of the time that \code{blog_dir} will point
to a local copy of a blog, and that the RSS file is at \verb{_site/index.xml},
it doesn't have to. \code{blog_dir} can also be a fully qualified path the RSS file
itself, or even the URL of an RSS file on the web, if you are using some other
blogging platform and dont want to copy the RSS file to your local filesystem.
}
\examples{
\dontrun{
## Using a local quarto blog,
## where index.xml is under _site/index.xml
b2n_compose("path/to/local/blog",
             from_email = "me@place.com",
             extra_subject = "[cool newsletter]")

## using a local index.xml that is not in a quarto blog, or a non-standard location
b2n_compose("path/to/local/blog/_out/index.xml",
            ...)

## using a url
bn2_compose("https://myblog.com/index.xml",
            ...)
}

}
