% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_metadata.R
\name{b2n_create_cache}
\alias{b2n_create_cache}
\title{create cache}
\usage{
b2n_create_cache(cache = "_blog2newsletter", r_script = "_blog2newsletter.R")
}
\arguments{
\item{cache}{the cache subdirectory to use}

\item{r_script}{the R script to use}
}
\value{
invisible
}
\description{
Creates the cache directory to be used by \code{blog2newsletter}.
}
