% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_email.R
\name{b2n_find_post}
\alias{b2n_find_post}
\title{find a post}
\usage{
b2n_find_post(blog_dir, post = "latest")
}
\arguments{
\item{blog_dir}{the location for the RSS}

\item{post}{which post to get}
}
\value{
tibble
}
\description{
Return either the \code{latest} post, or one matching a provided title.
}
