% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_metadata.R
\name{b2n_init}
\alias{b2n_init}
\title{create cache}
\usage{
b2n_init(r_script = "_blog2newsletter.R", cache = "_blog2newsletter")
}
\arguments{
\item{cache}{the subdirectory to use}
}
\value{
invisible
}
\description{
Creates the cache directory to be used by \code{blog2newsletter}.
}
