% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{b2n_run}
\alias{b2n_run}
\title{run full flow}
\usage{
b2n_run(run_file = here::here("_blog2newsletter.R"))
}
\arguments{
\item{run_file}{which file to run through}
}
\value{
invisible
}
\description{
Runs a workflow defined in \verb{_blog2newsletter.R}
}
