% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_metadata.R
\name{create_cache_directory}
\alias{create_cache_directory}
\title{create cache}
\usage{
create_cache_directory(cache = "_blog2newsletter")
}
\arguments{
\item{cache}{the subdirectory to use}
}
\value{
invisible
}
\description{
Creates the cache directory to be used by \code{blog2newsletter}.
}
