% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_email.R
\name{b2n_draft_email}
\alias{b2n_draft_email}
\title{creates draft in gmail}
\usage{
b2n_draft_email(post_email)
}
\arguments{
\item{post_email}{the post, created with \code{b2n_post_to_email}}
}
\description{
Given the composed email, sends the email to GMail so it can be viewed in the \emph{drafts} folder.
}
