% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_email.R
\name{b2n_send_newsletter}
\alias{b2n_send_newsletter}
\title{send it}
\usage{
b2n_send_newsletter(draft_email, cache = "_blog2newsletter", force = FALSE)
}
\arguments{
\item{draft_email}{email created by \code{b2n_compose}}

\item{cache}{the blog2newsletter cache location}

\item{force}{send it even if it was previously sent}
}
\description{
Send the drafted newsletter!
}
\seealso{
b2n_compose
}
